requirements.txt
pyautogui==0.9.53
pillow==9.2.0
opencv-python==4.7.0.72
numpy==1.24.3
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7